fix search with price and title and app is done

+++ lib/features/products/presentation/blocs/get_products/get_products_event.dart
@@ -6,7 +6,7 @@ class GetProductsEvent with _$GetProductsEvent {
   const factory GetProductsEvent.getMoreProducts() = GetMoreProducts;
   const factory GetProductsEvent.searchProducts({
     required String title,
-    required int minPrice,
-    required int maxPrice,
+    required double minPrice,
+    required double maxPrice,
   }) = SearchProducts;
 }
